templates: added full scaffolding for CRUD
Based on a model name, you can now create
a full set of CRUD operation boilerplate based on then
name of a model you pass to one of the following scripts:

  > gql:scaffold --model <model>
  > gql:scaffoldQuery --model <model>
  > gql:scaffoldMutation --model <model>

The last two, will generate only the Query scaffolding or
the Mutation scaffolding respectivley.

There is also the option to init a folder structure with an init
command.

TODO: write "gql:init" script