auth: initial ts migration & frontend config
Moved the front end to use a space's name, rather than an ID
in the query. The requested cards will be checked against the
space name in the backend which will help w/ API definitions in
the long run and help with client side data/routing.

Currently, url paths get passed to children components as variables
for each of the wrapped GraphQL components.