ui: global store hookup
The UI now utilizes a global store based off of React Context
and the useContext hook and the useReducer hook. With this method
of storing data now in place, it should smooth out the segregation
of different types of data, and help with hooking up the
frontend and backend, as well as data transmission to and fro.

TODO: update login/register side of app to use global store
TODO: update context to have localstore 'middleware'