be: updated graphql design pattern
The new design pattern for the backend is to take singular "inputs"
for queries and mutations. Each of these inputs types will be
an object that is passed to the system from the front end.

Purpose of this commit is to get everything setup, and then will
implement card features from here on out.